Setup man page generation from clap during build
Setup man page generation from clap during build so that we can
centralize the usage and in app docs within clap. This makes it much
easier to maintain rather than having to maintain two separate sets of
docs for usage that would naturally have a ton of overlap.

This relates to issue #239.

[changelog]
added: build time generation of man pages

<!-- ps-id: 5d43c4cd-f8e3-4b96-9d8a-7b62958ad285 -->