Add patch series support request_review_branch
Add patch series support request_review_branch so that users would be
able to either request review of a single patch or of a patch series.

Doing this was suprisingly involved.

It required completely changing how we were previously adding patch ids
to commits. Before we would effectively rebase the patch stack amending
the patch selected with a provided patch id if it didn't have one.

Now we take a bit of a different approach. Now we made it part of our
cherry picking functionality and as we cherry pick we can optionally add
a patch id to any commits missing them. This allows us to simply
add_patch_ids() near the beginning of any commands that need to make
sure the commits in the stack have patch ids.

Note: We have explicitly opted to call add_patch_ids() in these other
commands to avoid requiring users to use `gps c`. We deamed this
important as a number of users are using other Git tools in concert with
gps and we didn't want to force them to use `gps c`.

In order implement add_patch_ids() I first had to move the cherry
picking functionality over to ps/mod.rs so that I could add a
configuration parameter to it add_missing_patch_ids and when enabled
handle checking for patch ids as I cherry pick and generate and add any
missing ones. This makes sense to now be in ps/mod.rs as the cherry pick
methods are now aware of a patch stack concept, patch ids.

Once I had these functions I was able to rewrite the
request_review_branch() function to support individual patches or a
patch series. The general process is the same with a couple differences.

1. use new add_patch_ids() near the top
2. determine the associated branch in a different manner
3. cherry pick the patch series as ranged cherry pick

The last major piece of the puzzle was updating the CLI to accept a
String for the patch_index_or_range and then implelement a function to
parse that String into a PatchIndexRange type. This lets use provide the
patch index as <start-index>-<end-index>, e.g. 2-5.

[changelog]
added: patch series support to request-review-branch cmd

<!-- ps-id: f5726cb3-dbbb-4ee9-a50b-1c9482e8f7fe -->