Add alternating bg color to ls command
Add alternating bg in ls to alternate bg for patch series so that you
can visually see which rows are connected to each other. This is
especially important for being able to visually track the top most and
bottom most patch in a series from right to left to find their indexes.

This relates to issue #252.

[changelog]
added: alternating bg color to ls command

<!-- ps-id: 181190b5-8ae0-4f64-889d-1464fe90f880 -->