Remove the gps add command
Remove the gps add command because we are trying to reduce the surface
area of the commands and this command is simply a proxy command that
existed so people wouldn't have to leave the gps interface. However, we
are now shifting away from that mindset and more toward the mindset of
supporting switching seamlessly between git and gps commands.

[changelog]
removed: the add command

<!-- ps-id: cdfe7354-0f7b-4a32-80c9-62e390b458b1 -->