Rename the request-review-branch command to branch
Rename the request-review-branch command to branch because now that it
does everything branch did and more we don't need the distinction
between the two separate commands. Also it makes it clearer as to what
is actually going on.

[changelog]
changed: the request-review-branch command to just be branch
added: the command alias "b" to the new branch command

<!-- ps-id: dbcfdcc5-e26a-49da-9fdb-f4cf64270a33 -->