Remove the sync command
Remove the sync command because the request-review command now does
everything the sync command did now that it has the --no-post-sync-hook
and the --no-isolation-verification-hook options. This helps us collapse
the command surface area. Which will hopefully help new users be less
overwhelmed and be able to focus on learning the commands that they need
to.

[changelog]
removed: the sync command

<!-- ps-id: 2c556f81-91f2-4063-8a08-6af47e896325 -->