Add PatchIndexRangeBatch concept
Add PatchIndexRangeBatch concept to represent a batch or collection of
PatchIndexRange objects. This also houses the parsing logic to
constitute a PatchIndexRangeBatch from a string representation of the
form "1 3-5 7 9 10-13".

In the near future I plan to use this to implement batching into the
request-review command so that I can get rid of the batch-request-review
command and reduce the command surface area.

This relates to issue #255.

<!-- ps-id: 21328cb6-0e2a-49ed-8134-d0a205675f01 -->