Add batching support to the request-review command
Add batching support to the request-review command so that users would
be able to request review for a series of patches and patch series with
one go.

This also makes it so the request-review command supports the same
functionality as the batch-request-review command. Which will allow me
to now remove the batch-request-review command and reduce the surface
area of the commands so it isn't so overwhelming to new users.

This relates to issue #255.

[changelog]
added: batching support to the request-review command

<!-- ps-id: 524746ce-3a64-435d-8a7d-25149ad6be97 -->