Clean up error reporting for list command
Clean up error reporting for list command so that it is using all of the
std::error::Error features that were previously added.

Beyond that I switched to Box<dyn std::error::Error> for originating
errors because that is a different abstraction layer and we don't want
those lover level abstractions leaking into the consumers layer.

This in turn exposed that the GetRepoRootPathFailed(paths::PathsError)
case was no longer being used and therefore I removed that to clean it
up as well.

Then I switched the unhandled error printing to use the new
print_error_chain() utility that was recently added so that we can get
better error output.

[changelog]
changed: unhandled error output for the list command

<!-- ps-id: d634518a-5d45-468e-b512-866a237a4cf3 -->