Replace gpgme gpg signing with extern gpg usage
Replace gpgme gpg signing with external command gpg usage. This allows
us to be more inline with how Git implements these integrations. This
also enables us to eliminate the gpgme dependency, which in turn, should
eliminate a major blocker for build this on Windows. In general it also
simplifies building the project on all platforms because it eliminates a
build dependency.

As a side effect of this switch we now also respect the Git Config
gpg.program configuration.

This relates to issue #225.

[changelog]
changed: gpg signing to use external gpg command

<!-- ps-id: b6d6abd3-8ecc-4eb5-b30a-bc9112023e49 -->