Make gps list show played stack in rebase
Make gps list show played stack in rebase so that we can see the patches
in the stack that have already been played, including if they were
squashed, etc. and their associated meta data as well as the next todos
to be run as part of the rebase.

This relates to issues #273 and #280.

[changelog]
added: gps list support during rebases

<!-- ps-id: 920cba9f-581b-4c57-8b41-70c058e5d39f -->