Add behind count to list command
Add the behind count to the list command so that when someone does a gps
list they can see how far behind their stack is from the upstream.

The approach with this is to get the common ancestor between HEAD and
it's associated tracking branch and then to simply count the commits
between the upstream tracking branch and the common ancestor as the
count of those commits is how far behind your patch stack is from its
upstream.

This relates to issue #234.

[changelog]
added: behind count to the list (ls) command

<!-- ps-id: 34d0cebb-533a-42ba-8931-91ddf2e3dbfa -->