Add count_commits() function to the git module
Add count_commits() function to the git module so that I could easily
count the commits from one oid upto another oid. This is going to be
needed to support figuring out the number of commits we are behind when
running the gps list command.

The approach I took was simply to use the git2::revwalk() function to
build an iterator to walk the tree from the from_oid to the to_oid and
then consume the iterator counting the up the number of commits.

This relates to issue #234.

<!-- ps-id: 6cbab0ea-3efd-4791-9821-bfbc637aaed1 -->