Clean up sample hooks
Clean up sample hooks so that they are named consistently and in a
fashion that you can find the one you are looking for by name.

This naming scheme makes it easier to also understand what the proper
name of the hook should be.

Beyond that I made the default permissions of the hooks user executable
as that is also needed.

<!-- ps-id: c38370b7-0dbd-424d-b09e-213155d5839a -->