Look for namespaced signing program in git config
The program option in git config can live under [`gpg.<format>.program`](https://git-scm.com/docs/git-config#Documentation/git-config.txt-gpgltformatgtprogram) in
gitconfig, so we would need to respect that.

I moved fetching the program option to inside the format option match,
so we can use that to namespace the call to get the program
configuration, falling back to the legacy `gpg.program` config.

This small refactor is done since the ssh branch should also have an
optional custom program to run (the 1Password binary for example) and we
would want to follow the same general path with both methods.

https://github.com/uptech/git-ps-rs/issues/290

<!-- ps-id: 10064b8f-e14f-4f9e-aa06-97b8ac5b10f0 -->