Change integrate cmd to int from remote branch
Change integrate cmd to int from remote branch so that Github and
Bitbucket will detect integrations as merges. This now means that you
need to make sure that your patches are in sync with the remote prior to
integrating. So you will have to do an `request-review` to sync them up
prior if they are out of sync.

But this will result in GitHub and Bitbucket properly detecting the
integrations as merges once again.

<!-- ps-id: ebf0c061-8585-4bb3-b8eb-9cd5ffc829d0 -->