Add literal_ssh_key helper function
Add a function that returns a signing key if it is a literal value in
gitconfig, or None if it's not (would be a path).

This is done since with SSH signing, the user.signingKey in gitconfig
can either be a path to a file with the key, or a literal key (like with
gpg). See: https://git-scm.com/docs/git-config#Documentation/git-config.txt-usersigningKey

https://github.com/uptech/git-ps-rs/issues/290

<!-- ps-id: 46319555-2a7b-44a4-ac17-c1e1fd8cd72b -->