Add the `-q` option to the ssh-keygen cmd
Add the `-q` option to the ssh-keygen cmd to silence it so we don't see
the output informing us that we are "Signing data on standard input".
This doesn't match 100% with the way that Git proper is calling this
command anymore, because they don't use the `-q` option. Instead they
seem to execute the program ignoring it's standard output. Which maybe
we should do to make the interface match exactly the same. That way
other applications don't have to implement the `-q` switch that are
trying to standin, like the 1Password ssh key one.

<!-- ps-id: e9fcc740-16c6-4818-910d-8913ad58bdf0 -->