Rework ssh signing to match Git implementation
Rework ssh signing to match Git implementation so that users signing
with Git proper will also get signing simply by Git Patch Stack looking
at the Git config.

To accomplish this we simply had to replicate the implementation in Git
proper that uses ssh-keygen and how it behaves. Alon originally took the
first stab at this but it wasn't enough of a match to the implementation
in Git proper. For one it didn't handle the -U option being passed to
ssh-keygen when dealing with a literal key in the config. Beyond that it
also wasn't executing the ssh-keygen command the same way as it was
using stdin while the Git implementation was creating a temporary file
for the input.

So I shifted over to match the Git implementation. This in turn
eliminates the warning that ssh-keygen was producing on stderr about the
fact that we were signing content from stdin.

<!-- ps-id: a8ed768f-3b2e-49cd-a584-8e742508f799 -->