Fix bug with ssh signing for literal ssh keys
There was a bug that was preventing Git Patch Stack from successfully
signing commits/tags when the Git config is set up with a literal public
ssh key as the signing key in the configuration. This happens to also be
the configuration that 1Password uses with Git when signing commits with
1Password.

The reason it was failing was because the temporary directory,
containing the temporary file, containing the literal ssh public key was
being deleted prior to running the signing command. This of course
caused the signing command to execute out with an error. The deletion of
the temp directory is tied to the destruction of the TempDir object.

Therefore, to address this I moved the TempDir object up to a higher
scope that would allow it to live long enough for the signing command to
happen successfully.

[changelog]
fixed: ssh signing with literal ssh public key configured

<!-- ps-id: 3abe6d7a-5cc2-4b3a-831b-15b3cee19c9d -->