Add append subcommand
Add append subcommand so that users would be able to append a patch(es)
from their patch stack onto an existing patch/patch series branch.

gps append -n <branch-name> <patch-index-or-range>

This will facilitate workflows where you don't want to rewrite the
history of branches after you have shared them by pushing them up or
opening a related pull request. It does so by allowing users to continue
to add patches on their stack that are related to a patch/patch series
that was already requested for review and be able to add new patches on
top of that branch without leaving the patch stack, and doing it in a
way where git patch stack can still understand they are the same
patches.

[changelog]
added: append subcommand