Only show r in ls output when patch in remote branch
Previously the list command was showing the `r` state for patches that
didn't exist in the remote upstream tracking branch. This was a bug.
This corrects that by only showing the `r` and the following status
indicators if the patch is found in the remote upstream tracking branch.

This is a good fix in general. But specifically, it is aiding in the new
workflow I am currently using where we can `gps append` patches to a
branch that has already been pushed up. As this will now visually
indicate if a patch is in the local branch and the remote or just the
local. Letting you know if you need to push the changes up to the
remote.

[changelog]
fixed: bug showing r for patches in list command that weren't in the remote tracking branch